🐛fix : 마이페이지 거리, 시간 소수 첫번째 자리까지 반환

+++ backend/src/main/java/com/luckycookie/crewin/dto/MemberResponse.java
@@ -43,8 +43,8 @@ public class MemberResponse {
         private String nickname;
         private String email;
         private String name; // 사용자 이름
-        private int totalDistance;
-        private int totalTime;
+        private double totalDistance;
+        private double totalTime;
         private int totalAttendance;
     }
